backend/University-Semantic-Search-System/NLP-Pipeline/NLP Data/requirements.txt
# NLP Pipeline Requirements

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# PDF processing
PyMuPDF>=1.23.0  # fitz
pdfplumber>=0.10.0

# Text processing and NLP
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Vector search
faiss-cpu>=1.7.4  # Use faiss-gpu if CUDA available

# Progress bars
tqdm>=4.65.0

# Optional: for better performance
# faiss-gpu>=1.7.4  # Uncomment if using GPU
